requirements.txt
numpy
pandas
scipy
matplotlib
pyyaml
jupyter